# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Account:
    properties:
      email:
        description: Email address of the account holder
        example: alice@example.com
        type: string
      id:
        description: Unique identifier for the account
        example: "1"
        type: string
      name:
        description: Name of the account holder
        example: Alice Smith
        type: string
    type: object
  models.AccountErrorResponse:
    properties:
      code:
        description: HTTP status code representing the error
        example: 404
        type: integer
      error:
        description: Error message describing what went wrong
        example: Account not found
        type: string
    type: object
  models.AccountResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: Pointer to account, can be nil for empty responses
      message:
        description: Message describing the result of the operation
        example: Account retrieved successfully
        type: string
    type: object
  models.AccountsListResponse:
    properties:
      data:
        description: Slice of accounts
        items:
          $ref: '#/definitions/models.Account'
        type: array
      message:
        description: Message describing the result of the operation
        example: Accounts retrieved successfully
        type: string
      total:
        description: Total count of accounts
        example: 2
        type: integer
    type: object
...
  models.CreateAccountRequest:
    properties:
      email:
        description: Email address of the account holder
        example: alice@example.com
        type: string
      name:
        description: Name of the account holder
        example: Alice Smith
        type: string
    required:
    - email
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: HTTP status code
        example: 404
        type: integer
      error:
        example: Item not found
        type: string
    type: object
...
  models.UpdateAccountRequest:
    properties:
      email:
        description: Email address of the account holder
        example: alice@example.com
        type: string
      name:
        description: Name of the account holder
        example: Alice Smith
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is a sample CRUD API for inventory management system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory Management API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Retrieve all accounts from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountsListResponse'
      summary: Get all accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Add a new account to the system
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AccountErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AccountErrorResponse'
      summary: Delete an account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Retrieve a specific account by its unique identifier
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an existing account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AccountErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update an account
      tags:
      - accounts
...
swagger: "2.0"
...